Extract WMController diagnostics and trace surface into RuntimeDiagnosticsCoordinator
Move the ~42-function diagnostics cluster (runtime-state dump, reset,
restart-clearing-state, trace capture toggle and export, background
trace clip buffer, per-category runtime trace recorders,
viewport-mutation audit gating, window-decision debug snapshots) out of
WMController into a dedicated @MainActor coordinator owned by
WMController, following the existing handler pattern (weak controller
back-reference). WMController shrinks from 5,031 to 3,921 lines.

All callers address the coordinator directly via
controller.diagnostics.* — no routing signature (hotkeys via
CommandHandler, IPC debug endpoints, DebugBar, Settings Diagnostics
tab, AppRulesView, workspace bar tooltips) forced a forwarding shim, so
none were kept. RuntimeTraceCaptureStatus becomes a top-level type; the
trace export directory moves to
RuntimeDiagnosticsCoordinator.traceCaptureDirectory.

WMController changes beyond the move are minimal widenings the
coordinator needs: workspaceBarManager / debugBarManager /
hiddenWorkspaceBarMonitorIds become private(set),
relaunchCurrentApplication becomes internal, and a
clearExplicitWorkspaceMoveIntents() hook covers the one private table
the diagnostics reset clears.

No behavior change: trace file formats, pasteboard copies, command
palette debug actions, and nehirctl debug endpoints are byte-identical
code paths, only relocated. Test updates are the mechanical accessor
path (controller.diagnostics.*) only.
