--- .changeset/20260702002427-internal-preferred-frame-query-seam.md
@@ -1,6 +1,0 @@
----
-"nehir": none
-
----
-
-Internal: preferred-frame query seam
