+++ .changeset/20260702014850-redact-runtime-debug-dump-and-trace-export-file-.md
@@ -1,0 +1,6 @@
+---
+"nehir": patch
+
+---
+
+Redact runtime debug dump and trace-export file paths in runtime-debug logs (mark dump contents and export paths private; pasteboard/export flow unchanged). Internal: extract WMController diagnostics/trace surface into RuntimeDiagnosticsCoordinator and add a preferred-frame query seam.
